+++ apps/web/src/app/schools/classes/classes.module.ts
@@ -3,9 +3,11 @@ import { CommonModule } from '@angular/common';
 import { ClassesComponent } from './classes.component';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { SchoolsRoutingModule } from './classes-routing.module';
+import { ClassesService } from './classes.service';
 
 @NgModule({
   declarations: [ClassesComponent],
   imports: [SharedModule, CommonModule, SchoolsRoutingModule],
+  providers: [ClassesService],
 })
 export class ClassesModule {}